cleanup

+++ 2022/05/main.rs
@@ -3,9 +3,9 @@ use std::{collections::VecDeque, iter::*};
 #[derive(parse_display::Display, parse_display::FromStr, Debug, Clone, PartialEq, Eq, Hash)]
 #[display("move {num} from {from} to {to}")]
 struct Move {
-    num: i64,
-    from: i64,
-    to: i64,
+    num: usize,
+    from: usize,
+    to: usize,
 }
 
 type Parsed = (Vec<VecDeque<char>>, Vec<Move>);
@@ -15,8 +15,8 @@ fn part1(data: &Parsed) -> Answer {
     let mut stacks = data.0.clone();
     for mv in &data.1 {
         for _i in 0..mv.num {
-            let c = stacks[mv.from as usize - 1].pop_front().unwrap();
-            stacks[mv.to as usize - 1].push_front(c);
+            let c = stacks[mv.from - 1].pop_front().unwrap();
+            stacks[mv.to - 1].push_front(c);
         }
     }
     stacks.iter().map(|x| x.front().unwrap()).copied().collect()
@@ -27,11 +27,11 @@ fn part2(data: &Parsed) -> Answer {
     for mv in &data.1 {
         let mut tmp = VecDeque::new();
         for _i in 0..mv.num {
-            let c = stacks[mv.from as usize - 1].pop_front().unwrap();
+            let c = stacks[mv.from - 1].pop_front().unwrap();
             tmp.push_front(c);
         }
         for x in tmp {
-            stacks[mv.to as usize - 1].push_front(x);
+            stacks[mv.to - 1].push_front(x);
         }
     }
     stacks.iter().map(|x| x.front().unwrap()).copied().collect()
